refactor: split agent.rs at the seam it already had
Adding --log-file put it at 502 lines. `link_line` is the one pure
decision in a function that is otherwise entirely I/O, and it was already
carrying its own test module for exactly that reason — so that is where
it splits.
